add atoms: avatar, blueCheckbox and blueTextField
